profiling: symbolizer: adds llvm symbolizer (#2035)
Adds llvm symbolizer class which using a dynamic library and dlopen to
link to LLVM. This was necessary in order to eventually allow #1853 to
be implemented without relying on a subprocess.

Statically linking LLVM is not viable because Perfetto uses libc++ while
system LLVM is often built with libstdc++, leading to linker errors.