tp: don't intern tracks when parsing track descriptors (#2142)
This is no longer necessary as we've separated resolving tracks from
interning tracks.

I've ensured that all the changes to the test pass both before and
after: I had to make the changes because they were too brittle and
relying on track ids/parent ids. Instead we use track names (which
required me to add track names everywhere). Moreover, we also need to
emit events on every track as that's now required for a track to show up
in the track table now.