Add constructor for StringView from string_view (#1781)
This is useful in case you have a `string_view` and don't want to
allocate a copy or explicitly call the data/size constructor.

There is already an explicit constructor from `std::string`, but for
`string_view` I don't think there are any footguns and we should just
have a regular converting constructor.

Co-authored-by: Lalit Maganti <lalitm@google.com>