Extend frame boundary for per-frame metrics. (#2089)
In some cases when a blocking call crosses a frame boundary, the next
actual frame does not start till the blocking call ends. With the
current metric logic, the part of the blocking call outside the frame
boundary is not being considered for metric calculation, which is
incorrect.
Fixed the frame boundary logic to not consider the actual frame
boundary, but consider the start of the current frame to the start of
the next frame as a frame boundary.