Prepare to split chrome_track_event.proto into multiple files
An upcoming Chrome patch will split chrome_track_event.proto into
multiple files. This updates the build file so that patch won't fail
when it's synced in:

Generate chrome_track_event.descriptor from a variable in sources.gni.

Add the current directory to import_dirs so that imports in the split
files can be found.