ui: Fix lingering omnibox text bug
Currently, the omnibox text for a given mode likes to stick around
even after the mode has completed. This makes it confusing to e.g.
run the same command twice that prompts the user with a second omnibox
prompt, as the previous filter will still be present.

This patch just stores a single string to hold the omnibox text (as
opposed to one per mode) and clears it down whenever the mode changes.

Change-Id: Ic3f057ec016d71d9dadad05a65b52db5b630de09