ui: Fix lingering omnibox text bug (#2192)
Currently, the omnibox text for a given command likes to stick around
even after the command has completed. This makes it confusing when you
e.g. run the same prompt twice as the previous filter text will still be
present in the omnibox on the second attempt.

This patch switches to using \a single string to hold the omnibox text
(as opposed to one per mode) and clears it down whenever the mode
changes.