stdlib: Fix Wattson markers slice bug
Update the Wattson markers table to display only if the start marker is
defined. Consistent with how other slices are displayed in Perfetto, if
the end of slice is missing, display only the start with end trailing
off; if the end slice is missing, do not display the slice at all.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.wattson.'
Bug: 432505126
Signed-off-by: Samuel Wu <wusamuel@google.com>