tp: ensure that datatframe module points to correct vtab post finalize
Before we were pointing to the vtab in storage rather than the one in
the shared storage due to not invalidating the pointer in the vtab. By
not storing the pointer anymore, we workaround this issue.

The real fix is that we should drop/recreate the table to ensure all
vtabs are correctly invalidated. But that's a complex fix and will clash
with upcoming work to multi-thread trace processor