Templated helpers to decode messages with extensions:
GetFieldWithMetadata() and DecodeFieldWithMetadata() will find a
non-repeating field based on a generated FieldMetadata struct. These are
basically wrappers around ProtoDecoder::FindField() but add static
asserts that the FieldMetadata is from the expected message and has the
expected type.

CppMessageDecoder is a wrapper around ProtoDecoder that owns the
serialized message bytes. It can be used to convert a CppMessageObj from
the "cpp" plugin, which does not support extensions at all, to a pbzero
decoder.