tp: move tracking of trace_id for legacy Chrome events to slice (#2236)
Since now we multiplex slices in trace processor tracks, more than one
source_id can be associated with each track. This makes it not possible
to store this arg on the track. Instead store it on the slice.

Also intercept this at json export to strip it out and convert it to the
ID correctly.