tp: remove cost estimation for static table functions
Our cust estimates at the moment suck and bad estimates are worse than
no estimates at all. Not setting the estimate and row count means that
SQLite sets the maximum values and that means it will try and
materialize the data as much as possible. Thankfully that's exactly what
we want.