ui: add ability to specify rate-specific units in plugins
For things like power we may want to specify the rate-mode units
in watts instead of the autogenerated joules/sec style units in
order to present the user with something more meaningful.

if the plugin does not specify anything then unit/sec will continue
to be used for counter tracks in rate mode.