diff_tests: Include extra descriptors in serialize_test_trace command (#2271)
When a test fails, the framework tells you how to serialize the trace.
The command given by the framework is broken if the trace uses
extensions. For example, when the diff_test
`WindowManager:has_expected_rows` fails, the suggested command fails to
generate the trace.

This commit fixes the problem.