Reland: Move ChromeProcessDescriptor to an extension
Adds an extension range for Chrome to TrackDescriptor, and a
ChromeTrackDescriptor extension proto in protos/third_party/chromium.
This causes protos/third_party/chromium to diverge from Chrome's
base/tracing/protos, due to a dependency cycle between Perfetto and
Chrome. Chrome will need to import the changes to
protos/perfetto/trace/track_event and update base/tracing/protos in one
commit before the next sync from Chrome -> Perfetto.

Also adds an ExtensionsWireCompatibility test to make sure that
converting an inline field to an extension field like this is safe.

Now that ChromeProcessDescriptor is part of an extension, it's moved to
a different descriptor file which needs to be included when calling
`serialize_python_trace` for diff_tests. Every other `serialize_*_trace`
variant already takes a list of extension descriptors so this brings
Python in line with other functions.