tp: remove base::Hash and base::Hasher shims and push down FnvHash (#2286)
This CL pushes down explicitly use of base::FnvHash into all the calling
locations to make it easier to figure out which hash is being used at
any one time. This will allow us to switch the hash without accidentally
breaking backcompat in a place which expects a stable hash.