base: prepare murmurhash to actually be ready to use (#2303)
This CL does three things:

1) **Switch murmur implementation to faithful mmh64**

The custom implementation from DuckDB sadly was causing trivial
collisions in some real world traces. Use the actual implementation of
MMH64 which at least means we are depending on something reliable.

2) **Add MurmurHashCombiner to combine multiple values**

The combine function is taken from boost::hash_combine and we use it to
combine multiple mm64 hashes.

3) **Add MurmurHash to act as a drop in implementation to std::hash**

The class we're going to use in base::FlatHashMap to act as a hash. Uses
the mixing function for integers, mm64 for strings and std::hash for
everything else (though we should really add a absl like implementation
to allow arbitrary types to be hashed).