github: don't append 'github.run_id' to the build cache key. (#2320)
The 'github.run_id' is unique for each run, so when running the next
build
for the same job, the cache key is different. Instead of the most recent
cache,
one of the older caches is used (using 'restore-keys:').

We only update caches when building the main branch, so it is OK 
if more than one concurrent build is running, in this case one of the
writers
win and in the worst case slightly outdated (e.g. HEAD~1) build cache
will be saved.