stdlib: Remove check for dur > 0 on ii() tables
Remove check for dur > 0, which was originally in place to remove slices
with dur = 0. This is possible now with the update to ii().

The only dur > 0 checks that remain are for thread slices. Thread slices
can have a dur = -1 when the slices do not end (usually near the end
othe trace).

Bug: 422853790
Signed-off-by: Samuel Wu <wusamuel@google.com>