Add dmabuf_rss_stat trace event
This trace event reports the change in the amount of dma buffer memory
referenced by a process whenever a reference to a dma buffer is added or
removed. The total amount of dma buffer memory referenced by the process
is also reported.

This trace event is not upstream, and exists only in Android's GKI
kernels.

Bug: 424646834
Test: out/linux/traceconv text /tmp/dmabuf_rss_stat.trace | grep -B4 -A12 dmabuf_rss_stat