Call stack sampling: stack unwinding by kernel, skip symbolization.
This change introduces two new enhancements to stack unwinding and symbolization
to improve the sampling performance when callstack sampling is enabled.

Stack unwinding by kernel:
----------------------------
Currently Perfetto is using libunwinding which is runtime intensive and hence
reduces the sampling performance at higher frequencies. With this new feature,
unwinding happens in kernel-space instead of user-space. This is faster and
significantly improves the sampling performance. Kernel unwinding may not work properly
with JIT code and binaries that omit to store the frame pointer during function calls.

Skip symbolization:
--------------
Allows the user to skip the symbolization during the sample capture. Symbolization
needs to be done post capture using 'traceconv symbolize'. This option is recommended
to be used in combination with kernel callchain for performance.