docs: fix devloper tools stacked branch creation command
by default git new-branch creates the branch from main, instead of
the current branch. based on the comments in the docs, that does not
appear to be the intent on the example.

update the docs to create the 2nd branch using --current-parent so
that my-feature-2 is stacked on top of my-feature instead of main,
which appears to be the intent in the example.