ui: switch power rails from micro-watts to milli-watts
Most subsystems on a typical android device are in the milli-watt
range. Since these tracks come from android_power_rails packets in
the trace, targetting the typical range for an Android device is
likely to provide the best user experience.

In addition to changing the order of magnitude this assigns a proper
unit to the track, making use of the new rateUnit so that the displayed
unit appropriately swtiches between mJ and mW depending on if rate mode
is enabled or not.