tp: add unit and polarity to v2 metrics (#2362)
Adds the ability to specify units and polarity (i.e. whether a higher
or lower value is better) to v2 metrics.

This is exposed in the `TraceMetricV2Spec` and
`TraceMetricV2TemplateSpec`
protos and allows for richer, more machine-readable metric definitions
which can be used for automated analysis and visualization.

The documentation and tests have been updated to reflect these changes.

Fixes: 434666666