stdlib: Wattson updates for runtime improvement and readability (#2369)
When calculating the 2D dependency, the tables relied on finding both
the dependent CPU's frequency and dependent CPU's policy. The
combination of the CPU's frequency and CPU's policy can be uniquely
represented by the corresponding active power for that pair. This
simplifies the code and reduces runtime.

This change removes the device-specific table of minimum voting
configurations, and instead generically derives the minimum vote on
frequency/power from the pre-existing LUTs.

One side effect is that some outputs will have a slightly higher power
estimate, but this is intended and the proper implementation. In rare
instances where the static power LUT had a high value for the minimum
vote (occurs naturally due to noise and randomness in the calibration
process), the previous implementation would incorrectly choose lower
power values.

There are a few cleanups in this patchset for readability and
maintainability that also have the side effect of slightly faster
runtime.

In total, there is ~1-3% reduction in runtime of all Wattson test cases
(saves 100-300ms out of 9000ms on my setup).

tools/diff_test_trace_processor.py out/linux/trace_processor_shell
--name-filter '.*wattson.*'
Bug: 422853790
Signed-off-by: Samuel Wu <wusamuel@google.com>

---------

Signed-off-by: Samuel Wu <wusamuel@google.com>