ui: Display user friendly power rail names.
Pixels have several of their rails mapped to user frienly names in
trace processor which show up in the default track names. when we
moved to a dedicated PowerRails plugin these names disappeared in
favor of the raw name.

Since the goal of this plugin was user friendliness, bring back the
user friendly rail names, however without the "power.rails." this
time, both because it is redundant and because some of the track
names would not fit in the allocated space in the UI and would need
to be partially truncated.