stdlib: expose userfriendly power rail names. (#2388)
There exists a map of raw to userfriendly name in

https://github.com/google/perfetto/blob/0c893ed6bf6b42e3fee58daf3380d301c72550ed/src/trace_processor/importers/proto/android_probes_module.cc#L59-L120

We would like to expose these names to users of the stdlib without the
"power.rails." prefix which is part of the track name itself.

It is possible these user friendly name mappings should be moved into
the stdlib instead of mapping in trace processor, but it is possible
consumers of the power rails are referencing the track names in SQL
already so that will be left to a future change.

This PR mainly aims to create an API to get the userfrienly name that
can be depended upon by the ui regardless of if this mapping logic is
in trace processor or in the stdlib.

#Bug: 2249