github: use GCS to store 'buildtools/' cache. (#2370)
We use `tar` to create a single cache file from multiple files in
`buildtools/` directory, so we can rely on the atomicity of
`gcloud storage cp` command.
This prevents races during concurrent builds, ensuring only one writer
can successfully update the cache at a time.

When saving and restoring the cache, we use the `--keep-old-files` and
`--exclude-from=` `tar` flags to avoid overwriting files inside
`buildtools/` that are already stored in git.

We store a single, the most recent cache entry for each build type: we
update the cache only when building the main branch.

The content of `buildtools/` is rarely updated, so instead of uploading
the cache after each build, we compare the hashes of its base files to 
check if an update is needed.

An alternative was to calculate the checksum of the entire tar content
using
`tar --extract --to-stdout --file="$FILE" | sha256sum | awk '{print
$1}'`.
However, this was not feasible as it takes approximately 30 seconds to
compute on CI for the Android build (it contains multiple large binary
blobs).