tp: refactor proto modules to live in a separate struct (#2410)
This CL does a major refactoring across all of the proto parser and
module
to depend on a separate module context instead of the trace processor
context. This is the first step to stopping parser-local state from
using the context entirely for non-processor wide state.

For now, the struct is also stored inside the context but this will
change down the line to live in ProtoTraceReader.

While this change touches a lot of files, in practice, it's just quite
a mechanical change and there should be no behavioural change as a
result of this CL.