github: use GCS to store build caches.
To speed up the build cache save and restore, we now store it
in GCS, instead of the GitHub actions cache.

This PR adds a custom cache action that uses GCS to store the
data and also makes it possible to update the existing cache
entry. Updating the existing cache entry allows us to have a
single most recent build cache for each build type that is
updated on each commit in the main branch.

The action shell script is an updated version of a script used
in .github/actions/install-build-deps/action.yml.
The 'build-install-deps' caching will be migrated to use this
new cache action in the future PR.