ui: Allow state to be persisted when loading a trace using postmessage (#2343)
Traces loaded via postmessage are typically not sharable via permalink.
Instead, after the user clicks the share button, we can just show you
the URL to get back to that trace via the provider if a URL is supplied
via the postmessage args when the trace was first loaded. However this
doesn't allow for the ui state to be persisted.

This patch adds a feature where the ui state can be uploaded and the
hash injected back into the URL if a placeholder is provided, so that
the state may be restored when the trace is re-loaded.

This patch also contains a few opinionated cleanups:
- Made all fields of TraceSource readonly to indicate to users of
`openTrace` et al that the function will not mutate any of the fields.
- Split the permalink upload functions into two functions - one to
upload the trace blob and one to upload the ui state.