ui: Use a page container rather than the .pf-page class (#2457)
Currently every page must include the pf-page class in order to get
their page put in the right place in the grid, which is confusing and
makes it more difficult to use pages from within plugins.

This change uses a wrapper/container div and puts the page contents
inside the container instead. Pages can just put content in here and
it'll automatically scroll if it gets too high / wide, or they can set
their height (or width) to 100% and manage scrolling internally.