feat(ui): Add y-axis sharing for counter tracks
This change introduces the ability to share the Y-axis range between
multiple counter tracks in the Perfetto UI. This is useful for
comparing the values of different counters that share the same units.

A new field `y_axis_share_key` is added to the `CounterDescriptor`
proto. All counter tracks with the same `y_axis_share_key` and the
same parent track will share their Y-axis range.