Add a new low level grid component and use under various grid implementations.
This patch adds a new Grid component which takes the role of the presentation layer of all grid-like components. It basically just defines how a grid looks and defines some opinionated operations and actions you can do on the a grid.

Almost everywhere that has grids in the UI have now been changed to use this Grid as the underlying mechanism to reduce the amount of duplicated code, and to improve UI and UX consistency in the UI.

Change-Id: Ic48a99e32038596e58b814972ba81faaf6854418