tracebox: add elog if traced_perf is compiled out (#2483)
Minimal change to help users notice cases like
https://github.com/google/perfetto/issues/2481.

This is in the forking path, so it won't spam the "usage" invocations.
Doing something more complex like showing errors in UI is not worth it.
I'd rather spend time figuring out how to refactor traced_perf to not 
require libunwindstack outside of Android.