ui: Add theme CSS variables and use throughout widgets (#2485)
This PR adds various semantic theme colors and fonts to common.scss
specified using css variables. These are to be used throughout the UI
and can be customized to allow for theming of the UI such as dark mode,
and to enable the UI to fit more seamlessly when embedding into other
websites.

This PR also makes use of these themes across all widgets - allowing
them to be themeable. Support still needs to be extended across the
entire UI, but this will be done in upcoming PRs to help reduce the
scope of each one.