[chromium] Update chromium recorder to v2 (#2497)
This PR uses the new org.chromium.trace_metadata2 with a dedicated
DISCARD buffer.
This supports multiple tracing session and reduces risk of data loss.
The ds support was added on M139, which is now on Stable.

Drive-by: chromeConfig is removed in all ds except memoryInfra ones.
deprecated `clientPriority` is removed entirely. Only
`privacyFilteringEnabled` field is kept in specific ds where it's used.