docs: reword ftrace guide + add kernel track event docs (#2455)
Firstly, this patch rewrites the existing ftrace guide to be more to the
point and with fewer leaps of reasoning. I'm still not too happy with
the page (some instructions are duplicated too much across our
quickstarts, while other things are namedropped but not obvious), but
believe it's better than before.

Secondly, it adds a new section to the quickstart guide covering a few
example "kernel trackevent" macros for common cases like process-scoped
thread tracks. This includes examples of a declaration, callsite, and
the resulting UI. 

Lastly, it adds a reference page outlining all options for the
well-known fields for this "kernel trackevent" parsing.