CI: show more t ccache statistics.
We need stats only for the current build, not the aggregated
stats for all the previously cached builds.
When setting up the ccache, we zero stats for the cache, so
when the build is finished, the `ccache --show-stats` prints
statistics for the current build only.

Also add '--verbose' to show cache failures.