CI: setup ccache after the cache directory is restored. (#2521)
Otherwise, the previously zero-ed ccache stats are overwritten
with the aggregated stats for all the previously cached builds.

Also extract the setup logic to an action, since we want each
`ccache` invocation to access global env variables that we set
on the first step.

This is a fix for abc857abb468d31ad06d97098a8011de8309af6c