ui: Use theme color variables throughout the UI (#2514)
This patch makes use of color theme CSS variables consistently
throughout the UI, including for use on the canvas (excluding slice
colors, for now).

This patch also contains some stylistic changes, which were made because
a lot of colors are currently hard coded and don't necessarily fit
neatly into theme's color palette. Thus, some colors are tweaked
slightly, and some bits of the UI are swapped out for widgets. The
result, however, is a simpler and more visually consistent experience.