Fix Dumpstate diff test input trace python typing
We currently have typing errors for the input trace type in the
dumpstate diff tests since CSV is only a valid output format, not
input format.

To fix this and present a more flexible option for future diff tests,
implement a RawText type which can not only be use for battery stats
parsing, but also other dumpstate, logcat, and other plain-text files
we aim to create.