Use generated names for unknown Chrome processes and threads (#2526)
If a new entry is added to the chrome ProcessType (or ThreadType) enum
without adding its name to the list in ChromeStringLookup, uses the
generated enum name from the proto instead of kNullStringId. This
ensures traces from the field that only contain ids will always have
distinguishable names, if not always pretty ones.