tp: add trace type detection for standalone dumpstate txt file. (#2547)
We currently recognize battery stats check files as a dumpstate
file, but not the actual full dumpstate txt files that are present
in bug reports.

Add this which will open up the option for constructing simple dumpstate
diff tests with only a few lines of input each instead of having to
add new test artifacts to test/data for every corner case we wish
to add new tests for.