tp: fix empty line reading in ChunkedLineReader (#2548)
ChunkedLineReader currently will stop abruptly upon the first empty
line that is encountered. This change fixes that by removing the
assumption that a zero size line indicates a failure. Instead we
create a separate boolean in SpliceResult to indicate success/failure
independent of the result line size, which allows us to produce zero
size strings AND report success at the same time.

ChuckedLineReader is currently only used by standalone Logcat,
Battery Stats checkins, and standalone dumpstate.

Out of these, Dumpsate is the only one that will normally contain
blank lines, and thus is the only one impacted by this. For this
reason we update the diff test to include the first 6 lines of the
example dumpstate file instead of 3 so that it will include a blank
line. Prior to this change standalone dumpstate would bail out after
parsing the first 3 lines since the 4th line is typically blank.