tools: make git_set_parent error on invalid input branch
Setting the parent to a branch that does not exist can cause all
of the stack related tools to fail when updating or syncing if one
of the parents don't exist.

To help mitigate the impact of user errors, prevent the user from
setting an invalid branch vs erroring later.