stdlib: Fix logic with devfreq_calc
The intent of _use_devfreq_for_calc and _skip_devfreq_for_calc is to
distinguish between devices that require using devfreq for power
estimates and those that don't, respectively. Only one of the two tables
should have data, and the other one should be empty. Without this patch,
the logic isn't correct, and is exposed when there are multiple devices
in _use_devfreq, so fix the bug.

Bug: 422853790
Signed-off-by: Samuel Wu <wusamuel@google.com>