infra: Run presubmit diff against the merge-base of HEAD and origin/main (#2623)
When running presubmit checks, we currently check changed files between
HEAD and (by default) origin/main. However, if you're working on an
older branch and origin/main has moved on, this diff can include changes
that are not part of your branch.

This patch changes the presubmit checks to diff against the merge-base
(i.e. the closest common ancestor between origin/main and HEAD, so that
it includes only the files you've changed in your branch.