Add url params option for the record android trace script (#2624)
when we use the tools/record_android_trace, it will be good we can pass
the url params when open the trace in the browser. This is part of
change for this feature rquest
([2400](https://github.com/google/perfetto/issues/2400)) .

The default value are the same as before, so there is no logic changes.

**Test:**
run the script: ./gen_amalgamated_python_tools to generate the
record_android_trace and tested with an Android device trace, all works
as expected.

When pass the url-params with cmd: `tools/record_android_trace -c
perfetto.config --url-params test1 test2 test3`, the url-params is
append to the url correctly in the browser.