ui: Mask noisy UI elements in integration tests instead of plumbing testing mode through the UI (#2661)
Currently we have a testing mode boolean in AppImpl which changes the
behavior of the UI in a few ways, including hiding the query time taken
as this is noisy and can cause false positives in the regression /
screenshot tests.

In an effort to avoid having to plumb these things through, this PR
changes the way these fields are ignored, using the built-in playwright
feature that lets you mask off certain elements while taking a
screenshot, which is a much less invasive way of achieving the same
thing.