ui: refactor all UI commands to adhere to `com.namespace.CommandName` (#2666)
As commands now become an API that we expect people to start depending
on, it's time to have a formal and consistent way for commands to be
named. After discussion, we've settled on com.namespace.CommandName
similar to how plugins themselves are named.

We intentionally avoid putting the plugin name as it makes refactoring
of code in the core/plugins a pain (people become scared to do
refactoring etc.)