ui: Consistent UX for tabular data (DataGrid, PivotTable and SQLTable) (#2583)
## Overview

This patch aims to make the various different grid/table implementations
throughout the UI have a consistent UX in regards to the following
features:

- Column reordering via drag and drop
- Cell & column context menus
- Column sorting buttons
- Right alignment for numerical data
- Styling for null/missing values
- Sticky header
- A slot for aggregations in each header cell

## Implementation

In order to implement this without duplicating code, a new widget called
`Grid` has bee created which takes the role of the being a common
presentation layer for grid-like components. It subsumes the roles of
the current Table, CustomTable, and BasicTable, and It essentially
defines the look and feel of a grid but is entirely un-opinionated about
the data displayed or how it's loaded and manipulated.

This makes it an ideal base layer for the different types of grids in
the UI, while ensuring they all have a consistent UX.

Almost every grid like element in the the UI has now been changed to use
this Grid as the underlying mechanism to reduce the amount of duplicated
code, and to improve UI and UX consistency. The exceptions are the
virtually scrolling grids such as in the logs and ftrace details panels,
but they're next on the list

The list of tables that have been converted are:
- SQL Table
- Pivot Table
- DataGrid
- Event latency details panel
- Scroll details panel
- Scroll timeline details panel

## Screenshots
### SQL Table
<img width="987" height="417" alt="image"
src="https://github.com/user-attachments/assets/140acd5c-ce05-45bc-839d-d183509da11c"
/>

### Pivot Table
<img width="989" height="422" alt="image"
src="https://github.com/user-attachments/assets/f9c34ece-c187-40e7-8e44-e8291ddb3f1f"
/>