tp: etm: binary info error fix (#2587)
As a temporary fix for b/438916722.
ETM registers files with large sizes causing TraceBlobView size to
become truncated. This means we can not trust any of the below checks
for large files. So a solution is needed that is not too expensive
memory wise.

This PR converts these from errors to warnings and adds a TODO to find a
memory efficient fix later.