ui: Fix crash when loading traces larger than ~2GB (#2715)
Attempting to cache a trace larger than around 2GB is currently crashing
on when attempting to create an arraybuffer from a file for the sake of
caching. Caches can actually take a ReadableStream so we can just take a
stream from the file instead. Unfortunately it looks like we have a
separate problem in that chrome cannot cache files larger than 2GB
anyway.

Since caching is best effort anyway, this patch just makes it so that if
we fail to cache, we log it and don't convert the URL to contain the
cache UUID, but we ultimately silently ignore the issue and move on.

Fixes: https://buganizer.corp.google.com/issues/435023274