base: Introduce LockFreeTaskRunner
Introduce a new taskrunner class that is designed to replace
and sunset UnixTaskRunner. Thew new class avoids locks and also
avoids any allocations (*) for cases when there are < 512 tasks
outstanding (512 being the size of a slab). It does so by using
two Slabs and a freelist of size one like a graphics double buffer.

(*) It doesn't count for allocations caused by copying around
    tasks' std::function, if the bound task goes beyond the inline
    storage of std::function and requires heap.

Bug: b/394504259