ui: explicitly list commands we want to support running on startup (#2711)
We recently introduced a bunch of ways to run commands on startup (URL
params, settings) with potentially more coming up. We expect that people
will write these once and expect them to work ~forever (with any
breakages being treated as regressions).

Because of this, exposing the
full surface of all commands the UI can run is risky as it's very likely
that some of the plugins will evolve a lot faster and will expect to be
changes in API over time. Realistically, we don't care about actually
having all the commands, just a core list of ones which users will want
to be able to depend on.

Introduce a list of commands which we expect to be able to maintain in
the long term without breaking. This is a reasonably small list and one
I think should be pretty managable to maintain. Suggestions welcome if
we think this is too much and I'm happy to pare it down.

I've also added a test which enforces that we don't accidentally break
backcompat and a setting which allows power-users to turn off our
protection but with the understanding that we provide no back-compat
guarantees to them