Always use constexpr category registry with Find (#2735)
Profiling of Chrome shows that about 5% of the cycles used by Perfetto
tracing is in the TrackEventCategoryRegistry::StringEq method, which is
constexpr and documented to be intended for compile-time category
lookups. Since it's showing up in runtime traces, it's not always being
executed at compile time.

The root cause is that there are two global TrackEventCategoryRegistry
objects, `kConstExprCategoryRegistry` for compile-time lookups and
`kCategoryRegistry` for runtime lookups. In theory a single constexpr
object could be used for both purposes but due to
https://bugs.llvm.org/show_bug.cgi?id=51558 the object that includes a
pointer to runtime state data can't be constexpr. (See comments in
track_event_macros.h.)

StringEq() is only called from the Find() method, which is only called
in constexpr context. However one reference to it in
track_event_legacy.h uses the non-constexpr `kCategoryRegistry` variable
name, causing the compiler to execute the lookup at runtime.